Replace uniqueId with uuid
