+++ src/js/utils/general.js
@@ -60,6 +60,19 @@ export function setupTooltip(step) {
 }
 
 /**
+ * Create a unique id for steps, tours, modals, etc
+ * @return {string}
+ */
+export function uuid() {
+  let d = Date.now();
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+    const r = (d + Math.random() * 16) % 16 | 0;
+    d = Math.floor(d / 16);
+    return (c == 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+  });
+}
+
+/**
  * Generates a `Popper` instance from a set of base `attachTo` options
  * @param attachToOptions
  * @param {Step} step The step instance
